chore: refactor scheduleStorage types
